+++ nengo/tests/test_synapses.py
@@ -4,7 +4,7 @@ import pytest
 import nengo
 from nengo.processes import WhiteSignal
 from nengo.synapses import (
-    Alpha, filt, filtfilt, LinearFilter, Lowpass, SynapseParam, Triangle)
+    Alpha, LinearFilter, Lowpass, SynapseParam, Triangle)
 from nengo.utils.testing import allclose
 
 
@@ -34,7 +34,7 @@ def test_lowpass(Simulator, plt, seed):
     tau = 0.03
 
     t, x, yhat = run_synapse(Simulator, seed, Lowpass(tau), dt=dt)
-    y = filt(x, tau, dt=dt)
+    y = Lowpass(tau).filt(x, dt=dt)
 
     assert allclose(t, y, yhat, delay=dt, plt=plt)
 
@@ -45,7 +45,7 @@ def test_alpha(Simulator, plt, seed):
     num, den = [1], [tau**2, 2*tau, 1]
 
     t, x, yhat = run_synapse(Simulator, seed, Alpha(tau), dt=dt)
-    y = filt(x, LinearFilter(num, den), dt=dt)
+    y = LinearFilter(num, den).filt(x, dt=dt)
 
     assert allclose(t, y, yhat, delay=dt, atol=5e-6, plt=plt)
 
@@ -55,7 +55,7 @@ def test_triangle(Simulator, plt, seed):
     tau = 0.03
 
     t, x, ysim = run_synapse(Simulator, seed, Triangle(tau), dt=dt)
-    yfilt = filt(x, Triangle(tau), dt=dt)
+    yfilt = Triangle(tau).filt(x, dt=dt)
 
     # compare with convolved filter
     n_taps = int(round(tau / dt)) + 1
@@ -75,7 +75,7 @@ def test_decoders(Simulator, plt, seed):
     t, x, yhat = run_synapse(
         Simulator, seed, Lowpass(tau), dt=dt, n_neurons=100)
 
-    y = filt(x, tau, dt=dt)
+    y = Lowpass(tau).filt(x, dt=dt)
     assert allclose(t, y, yhat, delay=dt, plt=plt)
 
 
@@ -90,7 +90,7 @@ def test_linearfilter(Simulator, plt, seed):
 
     synapse = LinearFilter(num, den, analog=False)
     t, x, yhat = run_synapse(Simulator, seed, synapse, dt=dt)
-    y = filt(x, synapse, dt=dt)
+    y = synapse.filt(x, dt=dt)
 
     assert allclose(t, y, yhat, delay=dt, plt=plt)
 
@@ -119,7 +119,7 @@ def test_filt(plt, rng):
     k = 1. / tau * np.exp(-tk / tau)
     x = np.convolve(u, k, mode='full')[:nt]
 
-    y = filt(u, 0.1, dt=dt)
+    y = Lowpass(0.1).filt(u, dt=dt)
 
     plt.plot(t, x)
     plt.plot(t, y, '--')
@@ -136,9 +136,9 @@ def test_filtfilt(plt, rng):
     tau = 0.03
 
     u = rng.normal(size=nt)
-    x = filt(u, tau, dt=dt)
-    x = filt(x[::-1], tau, x0=x[-1], dt=dt)[::-1]
-    y = filtfilt(u, tau, dt=dt)
+    x = Lowpass(tau).filt(u, dt=dt)
+    x = Lowpass(tau).filt(x[::-1], y0=x[-1], dt=dt)[::-1]
+    y = Lowpass(tau).filtfilt(u, dt=dt)
 
     plt.plot(t, x)
     plt.plot(t, y, '--')
@@ -156,8 +156,8 @@ def test_lti_lowpass(rng, plt):
     lti = LinearFilter([1], [tau, 1])
 
     u = rng.normal(size=(nt, 10))
-    x = filt(u, tau, dt=dt)
-    y = filt(u, lti, dt=dt)
+    x = Lowpass(tau).filt(u, dt=dt)
+    y = lti.filt(u, dt=dt)
 
     plt.plot(t, x[:, 0], label="Lowpass")
     plt.plot(t, y[:, 0], label="LTI")